requirements.txt
numpy
matplotlib
numpy-stl
hexy
vpython